add chatroom subscription to infobar.js
